move parser module to request.rs
